LayoutTests/plugins/netscape-plugin-map-data-to-src.html

<html>
<head>
<script>
    if (window.layoutTestController)
        layoutTestController.dumpAsText();
</script>
<body style="margin:0px;overflow:hidden">
<object name="testPlugin" type="application/x-webkit-test-netscape" logSrc="true" data="blah">
</object>
<div>
    This tests that an object tag with a data attribute gets a src attribute as well.
</div>
<body>
</html>
