LayoutTests/plugins/netscape-plugin-map-data-to-src-expected.txt

CONSOLE MESSAGE: line 0: PLUGIN: src: blah

This tests that an object tag with a data attribute gets a src attribute as well.
